+++ botnet_test/Graph.h
@@ -36,19 +36,19 @@ public:
 
 	int Size() const {return m_size;}
 	size_t CommunityNumber() const {return m_communityNumber;};
-	double EdgeWeight(int i, int j) const;
-	bool IsCommunityEmpty(int comm) const;
+	double EdgeWeight(size_t i, size_t j) const;
+	bool IsCommunityEmpty(size_t comm) const;
 
 	double Modularity() const;
-	std::vector< std::vector<double> > GetModularitySubmatrix(const std::vector<int>& indices) const;
-	std::vector<double> GetCorrectionVector(const std::vector<int>& origCommInd, const std::vector<int>& destCommInd) const;
+	std::vector< std::vector<double> > GetModularitySubmatrix(const std::vector<size_t>& indices) const;
+	std::vector<double> GetCorrectionVector(const std::vector<size_t>& origCommInd, const std::vector<size_t>& destCommInd) const;
 	
-	void SetCommunities(const std::vector<int>& new_communities, int number = -1);
-	std::vector<int> Communities() const {return m_communities;};
-	std::vector<int> CommunityIndices(int comm) const;
+	void SetCommunities(const std::vector<size_t>& new_communities, int number = -1);
+	std::vector<size_t> Communities() const {return m_communities;};
+	std::vector<size_t> CommunityIndices(size_t comm) const;
 
-	void PerformSplit(int origin, int dest, const std::vector<int>& split_communities);
-	bool DeleteCommunityIfEmpty(int comm);
+	void PerformSplit(size_t origin, size_t dest, const std::vector<int>& split_communities);
+	bool DeleteCommunityIfEmpty(size_t comm);
 	void Print() const;
 	void PrintCommunity(const std::string& fileName) const;
 
@@ -57,12 +57,12 @@ private:
 	void FillModMatrix(const std::vector<int>& src, const std::vector<int>& dst, const std::vector<double>& weight);
 
 private:
-	int m_size;
+	size_t m_size;
 	double m_totalWeight;
 	size_t m_communityNumber;
 	bool m_isOriented;
 	std::vector<std::vector<double> > m_matrix;
 	std::vector<std::vector<double> > m_modMatrix;
-	std::vector<int> m_communities;
+	std::vector<size_t> m_communities;
 };
